refactor: 채팅 설정 변경
refactor: 채팅 소켓 메세지 위치를 RSideBar로 이동
