+++ src/components/rsidebar/RSidebar.tsx
@@ -3,10 +3,16 @@ import * as S from './RSidebar.style';
 import Todos from './todos/Todos';
 import ChatRoom from './chatRoom/ChatRoom';
 import { formatDateTime } from './utils/dateFormat';
+import { useSocket } from '@/socket/SocketContext';
+import { throttle } from 'lodash';
+import useChatStore from '@/stores/chat.store';
 
 type Tabs = '노트' | '할 일' | '채팅';
 
 const RSidebar = () => {
+  const socket = useSocket();
+  const setChatArray = useChatStore.getState().setChatArray;
+
   const [currentDateTime, setCurrentDateTime] =
     useState<string>(formatDateTime());
 
@@ -27,6 +33,43 @@ const RSidebar = () => {
       setSelectedTab(tab);
     }
   };
+
+  // socket
+  useEffect(() => {
+    if (!socket) {
+      return;
+    }
+
+    const handleReceiveChat = throttle((data) => {
+      setChatArray(data);
+      // console.log(data);
+    }, 300);
+
+    const handleNotice = throttle((data) => {
+      // console.log(data);
+      const noticeChat = {
+        nickname: 'notice',
+        message: data.message,
+        time: data.time,
+        imageUrl: '',
+      };
+      setChatArray(noticeChat);
+    }, 300);
+
+    socket.on('responseChat', (data) => {
+      // 채팅이 잘 보내졌나 확인
+      console.log(data);
+    });
+
+    socket.on('notice', handleNotice);
+    socket.on('receiveChat', handleReceiveChat);
+
+    return () => {
+      socket.off('recieveChat');
+      socket.off('responseChat');
+      socket.off('notice');
+    };
+  }, [socket]);
 
   return (
     <S.RSidebarStyle>
