+++ src/components/rsidebar/chatRoom/ChatRoom.tsx
@@ -7,7 +7,6 @@ import dayjs from 'dayjs';
 import { useSocket } from '@/socket/SocketContext';
 import useChatStore from '@/stores/chat.store';
 import { useAuthStore } from '@/stores/auth.store';
-import { throttle } from 'lodash';
 
 export default function ChatRoom() {
   const user = useAuthStore((state) => state.user);
@@ -33,12 +32,10 @@ export default function ChatRoom() {
       ...data,
       nickname: myNickname ? myNickname : '',
       time: dayjs().format('HH:mm'),
-    };
-    const newChat = {
-      ...reqData,
       imageUrl: user?.imageUrl ? user.imageUrl : '',
     };
-    setChatArray(newChat);
+
+    setChatArray(reqData);
     sendMessage(data.message);
   };
 
@@ -48,55 +45,16 @@ export default function ChatRoom() {
     }
   }, [chatArray]);
 
-  // 임시 socket 코드 작성
-  useEffect(() => {
-    if (!socket) {
-      return;
-    }
-
-    const handleReceiveChat = throttle((data) => {
-      setChatArray(data);
-      console.log(data);
-    }, 300);
-
-    const handleNotice = throttle((data) => {
-      console.log(data);
-      const noticeChat = {
-        nickname: 'notice',
-        message: data.message,
-        time: data.time,
-        imageUrl: '',
-      };
-      setChatArray(noticeChat);
-    }, 300);
-
-    socket.on('responseChat', (data) => {
-      // 채팅이 잘 보내졌나 확인
-      console.log(data);
-    });
-
-    socket.on('notice', handleNotice);
-    socket.on('receiveChat', handleReceiveChat);
-
-    return () => {
-      socket.off('recieveChat');
-      socket.off('responseChat');
-      socket.off('notice');
-    };
-  }, [socket, user]);
-
   const sendMessage = (message: string) => {
     if (message) {
       const payload = { message };
       socket?.emit('sendChat', payload);
     }
   };
-  // end
 
   return (
     <S.ChatRoomWrapper>
       <S.ChatArea>
-        {/* {추후 채팅들 배열로 가져와서 사용} */}
         {chatArray?.map((chatInfo, index) => (
           <Chatting
             key={index}
